refactor: extract components to files
